Add Delete Project Handler Permission Check Code (Green)

+++ TASKMANAGER.INFRASTRUCTURE/Handlers/Project/DeleteProjectHandler.cs
@@ -23,6 +23,10 @@ namespace TASKMANAGER.INFRASTRUCTURE.Handlers.Project
 
         public async Task<Unit> Handle(DeleteProjectCommand request, CancellationToken cancellationToken)
         {
+            var userPermissions = await _permissionsService.GetUserPermissions(request.UserId);
+            if (!userPermissions.SuperAdmin && !userPermissions.CanAddEditProject)
+                throw new TaskManagerException(ErrorCode.NoPermission);
+
             var project = await _projectRepository.GetByIdAsync(request.PublicId.ToString());
 
             await _projectRepository.DeleteAsync(project);
